--Added CardDetailsDto for getting the card details from the frontend-- --Minor adjustments to User and Product tables-- --Reworked the validatePayment method from OrderService--

+++ src/main/java/com/onlineshop/models/Product.java
@@ -30,7 +30,7 @@ public class Product {
     private String productSpecifications;
 
     @Column(name = "product_price")
-    private Integer productPrice;
+    private Float productPrice;
 
     @Column(name = "product_image")
     private String productImage;
